kb -> master

+++ Main.java
@@ -1,15 +1,48 @@
 import java.io.*;
 import java.util.*;
 public class Main {
-	public static void main(String[] args) throws IOException {
+	public static void main(String[] args) throws Exception {
+
+		System.out.println("function: ");
+		Scanner sc = new Scanner(System.in);
+
+		String choiseStr = sc.nextLine();
+		
+		
+		loop: while (true) {
+		switch(choiseStr) {
+		case "comp":
+			comp();
+			break;
+		case "decomp":
+			decomp();
+			break;
+		case "size":
+			size();
+			break;
+		case "equal":
+			equal();
+			break;
+		case "about":
+			about();
+			break;
+		case "exit":
+			break loop;
+		default:
+			System.out.println("no match");
+		}
+		}
+		
+		sc.close();
+	}
+	public static void comp() throws Exception{
 		LZ77 lz77 = new LZ77();
 		BitInput input = new BitInput(System.in);
-		BitOutput output = new BitOutput(System.out);
+		BitOutput output = new BitOutput(System.out);		
 		
 		Scanner sc = new Scanner(System.in);
+		System.out.print("Source file name: ");
 		String filename = sc.nextLine();
-		
-		
 		File file = new File(filename);
 		if (!file.exists()) {
 			System.out.println("File does not exist");
@@ -20,19 +53,91 @@ public class Main {
 			file.createNewFile();
 		}
 		FileInputStream fileInput = new FileInputStream(file);
-		FileOutputStream fileOutput = null;
-		
 		lz77.compress(fileInput, output);
+		FileOutputStream fileOutput = null;	
+		
 		lz77.decompress(input, fileOutput);
 		
-		byte[] bytes = filename.getBytes();
+		fileOutput = new FileOutputStream(file_1);
+		
+		String data = file_1.toString();
+		byte[] bytes = data.getBytes();
+		
+		
+		fileOutput.write(bytes);
+		
+		sc.close();		
+	}
+	public static void decomp() throws Exception{
+		LZ77 lz77 = new LZ77();
+		BitInput input = new BitInput(System.in);
+		BitOutput output = new BitOutput(System.out);	
+		Scanner sc = new Scanner(System.in);
+		
+		System.out.print("archive name: ");
+		String filename = sc.nextLine();
+		File file = new File(filename);
+		
+		if (!file.exists()) {
+			System.out.println("File does not exist");
+			return;
+		}
+		File file_1 = new File("decompressed_file.txt");
+		if(!file.exists()) {
+			file.createNewFile();
+		}
+		FileOutputStream fileOutput = null;	
+		
+		lz77.decompress(input, fileOutput);
+		
+		String data = file_1.toString();
+		byte[] bytes = data.getBytes();
 		
 		fileOutput = new FileOutputStream(file_1);
 		
 		fileOutput.write(bytes);
 		
-		fileOutput.flush();
 		sc.close();
+		output.close();	
+		fileOutput.close();
+	}
+	public static void size() throws Exception{
+		System.out.println("file name: ");
+		Scanner sc = new Scanner(System.in);
+		String filename;
+		filename = sc.nextLine();
+		Path path = Paths.get(filename);
+
+		try {
+			long bytes = Files.size(path);
+			System.out.println(String.format("%,d bytes", bytes));
+		}
+		catch (IOException e) {
+			e.printStackTrace();
+		}
+		sc.close();
+	}
+	public static void equal() throws Exception{
+		Scanner sc = new Scanner(System.in);
+		System.out.println("first file name: ");
+		String filename_1 = sc.nextLine();
+		System.out.println("second file name: ");
+		String filename_2 = sc.nextLine();
+		Path path1 = Paths.get(filename_1);
+		Path path2 = Paths.get(filename_2);
+		
+		byte[] f1 = Files.readAllBytes(path1);
+		byte[] f2 = Files.readAllBytes(path2);
+		if (Arrays.equals(f1,f2))
+			System.out.println("true");
+		else
+			System.out.println("false");
+		sc.close();
+	}
+	public static void about() {
+		System.out.println("Informācija par grupu:");
+		System.out.println("000RDB000 Filips Jānis Romāns 10.grupa");
+		System.out.println("201RDB207 Kārlis Baumanis 10. grupa");
 	}
 }
